backend/miniproject_app/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
djoser
djangorestframework-simplejwt
psycopg2-binary
pandas
numpy
scikit-learn
joblib
